More version updates suggested by Android Studio
Update Gradle support library to the latest stable version. Update
Gradle wrapper to the latest patch version too (required for some of
others updates). And update the Android test runners as well.

New test runners require AndroidX property to be enabled in the
properties file so add that as well.