New API of Secure Cell
Introduce new, interface-based API of Secure Cell. Current API has
a number of issues described in RFC 3.9, notably the run-time mode
setting and not-quite-secure support of String inputs which interferes
with passphrase API. That's why we introduce a new interface which
should rectify these issues.

The users will deal with *interfaces* of the base SecureCell class:
SecureCell.Seal, TokenProtect, ContextImprint. They will take into
account differences in inputs and outputs of these modes. They also
allow to introduce passphrase API which actually has the same API
after the initialization.

Using interfaces as opposed to classes allows us to hide unnecessary
implementation details and do not expose the users to a number of
classes that they don't actually need to interact with.

This is also a nice point to start paying more attention to JavaDocs,
with better descriptions of APIs and versioning.