Secure Cell passphrase API: RustThemis (#630)
* Add passphrase API of Secure Cell

Introduce new constructor "SecureCell::with_passphrase" which accepts
textual passphrases and returns a Secure Cell that handles them.

Currently only Seal mode supports passphrases so that's the only mode
that users are allowed to select.

The passphrase is accepted as AsRef<[u8]> to transparently allow use of
raw byte slices and other types, not limiting to AsRef<str> which could
be more appropriate (and more restrictive).

This commit does not update existing master key API for correct use of
symmetric keys and does not update API documentation as well. This will
be done separately.

* Integration test tool for passphrase API

Add "scell_seal_string_echo_pw" tool for Rust. Note that we need to add
it as a new binary in test tool crate. Testing infrastructure will build
and use it automatically based on the source file name.

* Simplify code sample for Secure Cell

Update the code sample like in other languages. Instead of a simple file
encryption tool (which is not a generally good use case for Secure Cell)
provide a showcase of Secure Cell APIs and how they impact the output.