Target Java 7 since that's what Android uses
We require Java 8 for building with embedded Gradle wrapper, however
Android still targets Java 7 by default at the moment. It is possible
to build for Java 8, but that changes the requirements to JVM. We'd
like to avoid cutting support unless we don't really need it.

Make it explicit that we're building for Java 7 for Android, and build
desktop Java code for Java 7 too. We use Java 7 features in the code
so we cannot build with Java 6. That train is now gone.

Also, change the annotation library to use the compatibility version
which supports Java 7.