Use JUnit 3.x pattern to test for exceptions
Unfortunately, we cannot use "assertThrows" which is so convenient.
First of all, we can't use lambdas from Java 8. However, even if we use
Runnable instead, JUnit wrapper on Android uses older JUnit which does
not have "assertThrows" support in any form.

Thus we fall back to ridiculously ugly, but working approach recommended
by JUnit authors [1].

[1]: https://github.com/junit-team/junit4/wiki/Exception-testing

Maybe when we bump the requirements to Java 8+ we could use better
testing library API. But not today.