Ignore tests that crash Android runners
These new tests are a pinata of crashes. They have discovered a serious
bug in JNI code which crashes the process on Android if a corrupted
Secure Cell is decrypted.

Disable these tests for now so that we can proceed with other updates.
They will be reenabled later, once the bug is fixed.