Passphrase support in JNI library
Add a new mode -- MODE_SEAL_PASSPHRASE -- to support Seal encryption
mode secured by passphrase. We do not introduce new API to JNI library
to keep ABI compatibility. Since the cost of doing JNI call is already
that high, it does not win us anything to use static mode-setting. It
makes sense to refactor Java API, but changing this JNI API does not
affect the users so it will be mostly a waste.

Note that Java constant is package-private. We don't want users to use
it with the old API because it will be extremely confusing. The old API
will accept strings, but it will encode them in UTF-16 by default which
is not what the new passphrase API does. Therefore we keep the constant
private. The old constants are getting deprecated too later.