Passphrase API implementation
Add a new package-private class SecureCellSealWithPassphrase which
implements SecureCell.Seal interface, but with passphrases instead
of symmetric keys.

The new class can be instantiated with SecureCell#SealWithPassphrase
methods that accept a variety of inputs. Normally the String interface
should be used to get compatible behavior. Other interfaces can be used
if you need some non-UTF-8 encoding for the passphrase, or if you're
doing something weird.

We need to store the passphrase bytes somewhere. A new helper class --
PassphraseBytes -- is added for that. It extends KeyBytes to implement
IKey interface and holds the sensitive data for us. It also hosts the
encodinng conversion routine. We cannot use the standard String#getBytes
because it replaces unknow characters with ?, leading to unexpected and
not really portable behavior.