Polyfill base64 decoding
It turns out that not all Android API levels support java.util.Base64
which appeared in Java 8. Since we build for Java 1.7, it's kinda wise
to not rely on that API. Android has its own API for decoding base64,
but we cannot use that for desktop Java.

So choose neither and provide a polyfill with is API-compatible with
java.util.Base64, implemented using an obscure Java API for processing
XML which include base64 support as well.