Use older version of Common Codec library
Unfortunately, we cannot use the latest version of the codec library
because some Android systems (e.g., API 24) already include an older
version of some of the Commons' classes. They are loaded before our
dependencies fetched by Gradle and break the party.

Investigation of Android source code shows that they're using what
appears to be version 1.2 of the library (the current version is 1.14).
Okay, so that limits us to the API provided by 1.2. Fine *sigh*