Configure Kotlin for testing
We are going to write Kotlin in JavaThemis to avoid adding new run-time
dependencies. Kotlin has great interop with Java due to this JVM thingie.
However, it will be nice to test our Java code with Kotlin to ensure that
our API works as intended.

Use the current latest stable version of Kotlin, configuruable via the
"kotlin_version" variable defined in the root project. There we also add
Maven Central repos where Kotlin pluggin and runtime are hosted.

Each subproject has Kotlin configured for it, even BoringSSL that does
not have any Kotlin code, but must have relevant targets as a
dependency. Note that desktop Java and Android use different plugin
because Kotlin compiles a little bit different on Android.

Kotlin runtime is required to run Kotlin code. Since we use Kotlin only
for tests, the runtime is added as test dependency. We build our code
for Java 1.7 so an appropriate runtime flavor is used.

Adding Kotlin also triggers a weird compilation error in Android's DEX
compiler as Kotlin runtime uses a lot of JetBrains annotations. This is
a known issue encountered by some project, and there is a workaround.