Secure Cell passphrase API: JavaThemis (#635)
* Passphrase support in JNI library

Add a new mode -- MODE_SEAL_PASSPHRASE -- to support Seal encryption
mode secured by passphrase. We do not introduce new API to JNI library
to keep ABI compatibility. Since the cost of doing JNI call is already
that high, it does not win us anything to use static mode-setting. It
makes sense to refactor Java API, but changing this JNI API does not
affect the users so it will be mostly a waste.

Note that Java constant is package-private. We don't want users to use
it with the old API because it will be extremely confusing. The old API
will accept strings, but it will encode them in UTF-16 by default which
is not what the new passphrase API does. Therefore we keep the constant
private. The old constants are getting deprecated too later.

* Passphrase API implementation

Add a new package-private class SecureCellSealWithPassphrase which
implements SecureCell.Seal interface, but with passphrases instead
of symmetric keys.

The new class can be instantiated with SecureCell#SealWithPassphrase
methods that accept a variety of inputs. Normally the String interface
should be used to get compatible behavior. Other interfaces can be used
if you need some non-UTF-8 encoding for the passphrase, or if you're
doing something weird.

We need to store the passphrase bytes somewhere. A new helper class --
PassphraseBytes -- is added for that. It extends KeyBytes to implement
IKey interface and holds the sensitive data for us. It also hosts the
encodinng conversion routine. We cannot use the standard String#getBytes
because it replaces unknow characters with ?, leading to unexpected and
not really portable behavior.

* Passphrase API tests

Once again, for the most part this is rewrite of Swift test suite into
Java. However, it has been augmented with some Java-specific tests to
verify that we handle the old 'passphrase-like' API in a sane way.