Destructuring accessors for SecureCellData in Kotlin
Kotlin supports tuples and destructuring assignment which results in
much nicer APIs with multiple return values. Implement special accessor
methods for SecureCellData to enable that.

These methods are not for general purpose use. They are expected to be
implicitly used in Kotlin when working with Token Protect mode, that's
why they also have @NotNull annotations -- as in Token Protect mode the
encrypted data and authentication token are always available.