Allow OutOfMemoryError in "swapTokenAndData" test
Here we use encrypted data as token which is essentially random.
Depending on how broken it turns out, Themis might genuinely believe
that this is a valid authentication token and proceed with allocation.
You get an exception if the allocation is bigger than JVM heap, which
it very well might be on emulated devices. Allow OutOfMemoryError to
happen in this test.