Always derive 32-byte (pre)keys from passphrases
Since we are going to use passphrase to derive a prekey, we should not
depend on the default encryption algorithm choice. Instead, always
derive full 32-byte prekey from passphrase with PBKDF2. After that
we will use Soter KDF to truncate the key to the length suitable for
the AES flavor in use. soter_alg_key_length() verification will also
be performed at that point to ensure that Secure Cell header is not
corrupted.