Apply Soter KDF after PBKDF2 for passphrases
Instead of using PBKDF2 output directly, pass the resulting prekey
through an additional round of Soter KDF to associate the encryption key
with user-provided associated context as well as the encrypted message
length, similar to how symmetric key API does this.

This changes the encryption algorith, but it's fine to do this before
public relese. This is the only chance we have to do this update without
it becoming a massive backward-compatibility pain.

The change slightly improves security of the encryption and makes the
algorithms a bit more compatible and similar in structure. However,
this does not change the API or data format in the slightest.