Destructuring accessors for SecureCellData in Kotlin (#638)
* Destructuring accessors for SecureCellData in Kotlin

Kotlin supports tuples and destructuring assignment which results in
much nicer APIs with multiple return values. Implement special accessor
methods for SecureCellData to enable that.

These methods are not for general purpose use. They are expected to be
implicitly used in Kotlin when working with Token Protect mode, that's
why they also have @NotNull annotations -- as in Token Protect mode the
encrypted data and authentication token are always available.

* Replace Arrays.copyOf() calls

@NotNull annotations remove a layer of optionals, so instead of
ByteArray? results we are now dealing with just ByteArray, and
that means we can call copyOf() method directly instead of using
java.utils.Arrays.copyOf(). Replace the calls to make linter happy.