Build ObjCThemis for generic iOS devices
Verify builds of ObjCThemis for iOS not only for the simulator
(i386, x86_64) but for "generic device" as well, using architectures
required by iOS (ARM64, ARMv7). This is not equivalent to building
an "archived" build of application -- which requires provisioning
profiles and all that other stuff -- and it does not verify bitcode,
but it's a good stopgap measure against OpenSSL maintainers forgetting
to include ARM builds into their precompiled framework binaries.

With Carthage we have an Xcode project so we can build it directly.
CocoaPods does not generate standalone project so we use one of the
examples and build the pods library. We cannot build the example for
"generic device" because of provisioning profiles, etc., but we can
build the framework alone which is enough to check that all symbols
required by ObjCThemis are actually present.