Claim official iPadOS support (#641)
iPadOS is actually a flavor of iOS and have unofficially supported
targeting iPads for quite a while (it technically worked). Check the
"iPad" check boxes is Xcode projects and run half of the test suite on
iPad simulators. Now we can claim with some confidence that we support
iPads and commit to resolving issues with that platform if they arise.