Ignore panicking corruption test in RustThemis
This test triggers a panic in "Vec::reserve" call in Token Protect
implementation on 32-bit machines. Corrupted input data is not detected
right away by Themis and we try to allocate more than 2 GB of memory.
It fails on 32-bit systems and Rust panics.

There is a nightly-only experimental API -- Vec::try_reserve -- but we
need to support stable Rust.

We can't do much about this issue other than assume a safe allocation
interval which is only an assumption. So for now let's panic, and
disable this test on 32-bit machines so that the test suite is still
useful.