Improve Java library path reporting
Sometimes "jnidir" gets duplicate slashes in it if the user sets PREFIX
to something like "/usr/". We check Java library paths by textual
inclusion so duplicate slashes can fail to match. Normalize the path
by cding into that directory and printing its absolute path. (This is
the most portable way to normalize a directory path.)

Furthermore, the "grep -vq" hack is incorrect and matches all inputs.
Do this check in a different way: look for the "jnidir" value in the
library paths and invert the result in shell conditional.