Make selectable languages in integration tests
When testing on legacy OS we know in advance that some tests for unsupported modules will fail. We want to make the testing behaviour meet real expectations so it is convenient to have configurable set of modules to test.

It was already implemented in the `tests_generator.py` script as environment variable `GENERATE_LANGUAGES`.

In this commit:

* the same behaviour was added to the `check_keygen.sh` script.
* `GENERATE_LANGUAGES` was renamed to `THEMIS_TEST_LANGUAGES`