Pin Emscripten toolchain to 1.39.16-fastcomp
Starting with Emscripten 1.39.17-fastcomp, the compiler is now printing
warnings, urging to migrate to the upstream flavor:

    emcc: error: the fastomp compiler is deprecated.  Please switch
    to the upstream llvm backend as soon as possible and open issues
    if you have trouble doing so [-Wfastcomp] [-Werror]

Unfortunately, we cannot migrate at the moment because the upstream
compiler is not able to build and link everything. While we're
investigating the issues, we need this warning gone -- we keep
a non-negotiable "treat warnings as errors" policy.

We cannot suppress this particular warning with "-Wno-fastcomp" because
it affects the BoringSSL build as well. They have "-Wall" added *after*
all custom flags so we are not able to suppress this warning there.

Therefore, for the time being, let's pin Emscripten toolchain to the
latest good version that does not emit warnings. I got the message,
I'll work on supporting the upstream flavor in the next release.