Support Node.js v14
This is the version which is going to turn into LTS in October 2020.
It's the current stable release. LTS is recommended for general use,
but some packages already deliver the stable version (e.g., Homebrew).

There are probably some deprecations that we need to take care of, but
otherwise we can squeese this support claim into the release window
for JsThemis 0.13 and WasmThemis 0.13.