Update and clean up changelog (#660)
* Add missings PRs since Themis 0.12 to the changelog

There are a lot of minor changes that can be grouped up as well as some
important changes that should be noted there.

* Fix typos in PR links

Many links are missing a closing parenthesis, add it. Markdown makes
typos like this too easy to do.

* Improve new and removed platforms list

Make the infrastructure part of the changelog easier to comprehend
by splitting out platform support changes into sublists.

* Update release highlights

Aside from the passphrase API, there are a couple of other imrovements
that are worth noting, such as updated documentation server. (The URL
still leads to the old one, we'll update it once it is deployed.)