Pin emsdk to avoid regression
The current HEAD of emsdk repository includes a change that causes the
builds to fail with an error:

    Adding directories to PATH:
    PATH += /home/user/emsdk
    PATH += /home/user/emsdk/fastcomp/emscripten
    PATH += /home/user/emsdk/node/12.18.1_64bit/bin

    Setting environment variables:
    EMSDK = /home/user/emsdk
    EM_CONFIG = /home/user/emsdk/.emscripten
    EM_CACHE = /home/user/emsdk/fastcomp/emscripten/cache
    EMSDK_NODE = /home/user/emsdk/node/12.18.1_64bit/bin/node
    /home/user/emsdk/emsdk_env.sh: line 26: /home/user/emsdk/emsdk_set_env.sh: No such file or directory

The change [1] has optimized environment variable setup. Apparently,
it has issues. To avoid breaking our builds, roll back to the previous
known good state for the time being. The issue has been reported to
Emscripten [2], but it's not clear when it's going to be reverted and
fixed, so we'd rather not depend on that right now, when Themis release
is brewing.

[1]: https://github.com/emscripten-core/emsdk/pull/539
[2]: https://github.com/emscripten-core/emsdk/issues/540