Stable boilerplate about autogenerated code (#664)
Recently released Bindgen 0.54.1 includes a change [1] which adds
Bindgen version into generated files by default. Since we keep the
generated files checked in and verify that it does not change, this
feature breaks our build by making it nonreproducible.

That is, the comment now reads

    /* automatically generated by rust-bindgen 0.54.1 */

and will change with each new version of Bindgen.

[1]: https://github.com/rust-lang/rust-bindgen/pull/1814

Use an option to disable the autogenerated comment completely and
replace it with our own warning about manually editing the file.

This means that we require Bindgen 0.54.1+ (for developing RustThemis,
not for building software using it). RustThemis developers are likely to
use the latest version (duh), but point it out in the changelog anyway.