Update Android artifact path for publishing
Introducing desktop Java build has moved Android project into a
subproject, not a top-level project. This changes the artifact name
which is based on the directory where build.gradle resides. It changes
from "projects" (repo directory name in the Docker container we use)
to "android" (that is "src/wrappers/themis/android").

Also, Gradle seems to be really confused with $buildDir when subprojects
are used so replace it with just "build" (relative to "build.gradle").