Keep on-demand configuration enabled (#671)
As previously noted, having this option disabled makes it so that
Android SDK is required to build desktop Java targets (":desktop...).
Unfortunately, this breaks Buildbot checks which does not have
Android SDK installed it is build environments.

However, we also need to disable this option to work around an issue
in Bintray publishing.

Since publishing to Bintray is a manual action and build checks are
automatic, let's favor the automation. We will have to manually disable
the option when making AndroidThemis release, but keep it enabled for
regular builds to pass. Hopefully, we'll fix the issue some day.